fix: moved animation to right component; chore: reverted imports from package.json;
