+++ packages/shared/package.json
@@ -130,8 +130,5 @@
   "sideEffects": [
     "./src/lib/lazysizesImport.ts",
     "*.css"
-  ],
-  "imports": {
-    "#*": "./src/*"
-  }
+  ]
 }
