+++ packages/shared/src/components/cards/ActionsButtons/index.ts
@@ -1,0 +1,4 @@
+import ActionButtons, { ActionButtonsProps } from './ActionButtons';
+
+export type { ActionButtonsProps };
+export default ActionButtons;
