--- packages/shared/src/components/cards/list/ActionButtons/index.ts
@@ -1,3 +1,0 @@
-import ActionButtons from './ActionButtons';
-
-export default ActionButtons;
