+++ packages/webapp/next.config.js
@@ -3,6 +3,7 @@ const withPWA = require('next-pwa');
 const withTM = require('next-transpile-modules')(['@dailydotdev/shared']);
 const { version } = require('../extension/package.json');
 const runtimeCaching = require('./cache');
+const path = require('path');
 
 const withBundleAnalyzer = require('@next/bundle-analyzer')({
   enabled: process.env.ANALYZE === 'true',
@@ -10,10 +11,10 @@ const withBundleAnalyzer = require('@next/bundle-analyzer')({
 
 const securityHeaders = [
   {
-      key: 'X-Frame-Options',
-      value: 'DENY'
-  }
-]
+    key: 'X-Frame-Options',
+    value: 'DENY',
+  },
+];
 
 module.exports = withTM(
   withPWA({
@@ -21,102 +22,102 @@ module.exports = withTM(
       dest: 'public',
       disable: process.env.NODE_ENV === 'development',
       runtimeCaching,
-      buildExcludes: [/react-syntax-highlighter|reactSyntaxHighlighter/]
+      buildExcludes: [/react-syntax-highlighter|reactSyntaxHighlighter/],
     },
-      ...withBundleAnalyzer({
-        i18n: {
-          locales: ['en'],
-          defaultLocale: 'en',
-        },
-        compiler: {
-          reactRemoveProperties: { properties: ['^data-testid$'] },
-        },
-        webpack5: true,
-        webpack: (config, { dev, isServer }) => {
-          config.module.rules.push({
-            test: /\.svg$/i,
-            issuer: /\.[jt]sx?$/,
-            use: [
-              {
-                loader: '@svgr/webpack',
-                options: {
-                  icon: true,
-                  svgo: true,
-                  replaceAttrValues: {
-                    '#fff': 'currentcolor',
-                    '#FFF': 'currentcolor',
-                    '#FFFFFF': 'currentcolor',
-                  },
-                  svgProps: {
-                    className: 'icon',
-                  },
+    ...withBundleAnalyzer({
+      i18n: {
+        locales: ['en'],
+        defaultLocale: 'en',
+      },
+      compiler: {
+        reactRemoveProperties: { properties: ['^data-testid$'] },
+      },
+      webpack5: true,
+      webpack: (config, { dev, isServer, dir, defaultLoaders }) => {
+        config.module.rules.push({
+          test: /\.svg$/i,
+          issuer: /\.[jt]sx?$/,
+          use: [
+            {
+              loader: '@svgr/webpack',
+              options: {
+                icon: true,
+                svgo: true,
+                replaceAttrValues: {
+                  '#fff': 'currentcolor',
+                  '#FFF': 'currentcolor',
+                  '#FFFFFF': 'currentcolor',
+                },
+                svgProps: {
+                  className: 'icon',
                 },
               },
+            },
+          ],
+        });
+        config.module.rules.push({
+          test: /\.m?js/,
+          resolve: {
+            fullySpecified: false,
+          },
+        });
+        return config;
+      },
+      env: {
+        CURRENT_VERSION: version,
+      },
+      rewrites: () => [
+        {
+          source: '/api/:path*',
+          destination: `${process.env.NEXT_PUBLIC_API_URL}/:path*`,
+        },
+        {
+          source: '/search',
+          destination: '/search/:provider',
+          has: [
+            {
+              type: 'query',
+              key: 'provider',
+            },
+          ],
+        },
+        {
+          source: '/search',
+          destination: '/search/posts',
+        },
+      ],
+      redirects: () => {
+        return [
+          {
+            source: '/posts/finder',
+            destination: '/search?provider=posts',
+            permanent: false,
+          },
+          {
+            source: '/signup',
+            destination: '/onboarding',
+            permanent: false,
+          },
+        ];
+      },
+      headers: async () => {
+        return [
+          {
+            source: '/:path*',
+            headers: [
+              ...securityHeaders,
+              {
+                key: 'X-Recruiting',
+                value:
+                  'We are hiring! Check https://daily.dev/careers for more info!',
+              },
             ],
-          });
-          config.module.rules.push({
-            test: /\.m?js/,
-            resolve: {
-              fullySpecified: false,
-            },
-          });
-
-          return config;
-        },
-        env: {
-          CURRENT_VERSION: version,
-        },
-        rewrites: () => [
-          {
-            source: '/api/:path*',
-            destination: `${process.env.NEXT_PUBLIC_API_URL}/:path*`,
           },
-          {
-            source: '/search',
-            destination: '/search/:provider',
-            has: [
-              {
-                type: 'query',
-                key: 'provider'
-              }
-            ]
-          },
-          {
-            source: '/search',
-            destination: '/search/posts'
-          }
-        ],
-        redirects: () => {
-          return [
-            {
-              source: '/posts/finder',
-              destination: '/search?provider=posts',
-              permanent: false
-            },
-            {
-              source: '/signup',
-              destination: '/onboarding',
-              permanent: false
-            }
-          ]
-        },
-        headers: async () => {
-          return [
-              {
-                  source: '/:path*',
-                  headers: [
-                    ...securityHeaders,
-                    {
-                      key: 'X-Recruiting',
-                      value: 'We are hiring! Check https://daily.dev/careers for more info!'
-                    },
-                  ]
-              }
-          ]
-        },
-        poweredByHeader: false,
-        reactStrictMode: false,
-        productionBrowserSourceMaps: process.env.SOURCE_MAPS === 'true',
-      }),
+        ];
+      },
+      poweredByHeader: false,
+      reactStrictMode: false,
+      productionBrowserSourceMaps: process.env.SOURCE_MAPS === 'true',
+    }),
   }),
 );
